refactor: put report assembly beside the writer and keep the journal gate honest
Assembling a report is a property of the artifact rather than of the run: it
takes cells already on disk and calls no provider. It belongs in emit.py, beside
the code that writes the three files, not in the module that drives the matrix.

Moving it would have blinded check_recording_harness_journalled, which decides
"is this a recording driver" by looking for a report constructor and would then
have found none in runner.py at all. A named assembler now counts alongside the
constructors it wraps, so a driver reaching a report THROUGH a helper is still
held to journalling the cells that paid for it, and no harness can leave the
gate by moving one call one module sideways.

The unresolved-claims caveat moves to models.py with its two siblings, because a
report's caveats are one vocabulary and splitting them across whichever module
raises each one is how two of them come to contradict each other.

Also fixes a caveat that could lie: a re-score without --repair-spend-from
inherited the sentence saying the token column had been rebuilt, while emitting
the raw journalled figures. The journal holds raw numbers and the repair is
applied at scoring time, so that caveat is derived from this run's flags and is
dropped rather than carried.
