+++ evals/recursion_depth/runner.py
@@ -42,13 +42,13 @@ from evals.errors import (
     HarnessProviderMissingError,
     OracleUnusableError,
     RecursionDepthGateUnbuildableError,
-    RecursionDepthNoCellsMeasuredError,
     RecursionDepthPlannerSubstitutedError,
     RecursionDepthSessionCeilingError,
 )
 from evals.harness.journal import RecordedCells
 from evals.harness.workspace import CellWorkspace
 from evals.recursion_depth.claims import requirement_ids_of
+from evals.recursion_depth.emit import assemble_report, derived_caveats
 from evals.recursion_depth.execute import LeafOutcome, leaf_task, run_leaf
 from evals.recursion_depth.forecast import estimate_sessions
 from evals.recursion_depth.gate import (
@@ -85,11 +85,6 @@ from evals.recursion_depth.models import (
 )
 from evals.recursion_depth.oracle import run_oracle
 from evals.recursion_depth.planner import PlanningSpend, TreePlanner
-from evals.recursion_depth.score import (
-    achieved_depth_histogram,
-    curve_by_achieved_depth,
-    curve_by_depth_cap,
-)
 from evals.recursion_depth.session import (
     SessionLimits,
     SweepDeps,
@@ -119,7 +114,6 @@ from synthorg.observability.events.evals import (
     EVALS_RECURSION_CELL_RECORDED,
     EVALS_RECURSION_CELL_RESTARTED,
     EVALS_RECURSION_CELL_UNAVAILABLE,
-    EVALS_RECURSION_NO_CELLS,
     EVALS_RECURSION_PLAN_RETRIED,
     EVALS_RECURSION_QUOTA_EXHAUSTED,
     EVALS_RECURSION_SESSION_CEILING,
@@ -170,14 +164,6 @@ _CEILING_CAVEAT: str = (
     "The sweep stopped early on its session ceiling, so the depths and "
     "repetitions the manifest asked for are not all present. Read the cell "
     "list, not the manifest, for what was actually measured."
-)
-
-_UNRESOLVED_CLAIMS_CAVEAT: str = (
-    "{dropped} planner claim(s) named no requirement this specification "
-    "defines and were dropped before scoring. A handful is one planner "
-    "inventing a requirement; a large share means the criterion template and "
-    "the id pattern have drifted apart, which deflates both halves of the "
-    "survival ratio and reads on the chart like a gate that does not help."
 )
 
 _QUOTA_CAVEAT: str = (
@@ -616,72 +602,6 @@ async def run_sweep(
         cells=cells,
         caveats=caveats,
         planned_cells=len(planned),
-    )
-
-
-def derived_caveats(cells: Sequence[CellRecord]) -> list[str]:
-    """The caveats a set of recorded cells implies on its own.
-
-    Separated from the assembler so the caller owns the final list. A re-score
-    carries the original report's caveats forward verbatim, which is the only
-    way the run-state ones survive (the ceiling and quota caveats are appended
-    while the loop runs and are recoverable from nowhere else), and an
-    assembler that appended its own on top would emit this line twice.
-
-    Returns:
-        The caveats these cells imply, which may be none.
-    """
-    dropped = sum(unit.unresolved_claims for cell in cells for unit in cell.units)
-    return [_UNRESOLVED_CLAIMS_CAVEAT.format(dropped=dropped)] if dropped else []
-
-
-def assemble_report(
-    *,
-    provenance: Provenance,
-    cells: Sequence[CellRecord],
-    caveats: Sequence[str],
-    planned_cells: int,
-) -> RecursionDepthReport:
-    """Build the report from cells that are already measured and journalled.
-
-    The single owner of how a report is shaped, so the recorder and the
-    re-score cannot drift apart while both keep producing a valid report.
-
-    Args:
-        provenance: What the sweep was measured against. Supplies the
-            requirement count both curves divide by.
-        cells: Every recorded cell, measured or unavailable.
-        caveats: The final list, already complete. Nothing is appended here.
-        planned_cells: How many cells the matrix asked for, for the log line
-            below. A re-score passes the recorded count, since everything on
-            disk was by construction planned.
-
-    Returns:
-        The assembled report.
-
-    Raises:
-        RecursionDepthNoCellsMeasuredError: Every run is unavailable.
-    """
-    measured = tuple(record for record in cells if record.achieved_depth is not None)
-    if not measured:
-        msg = (
-            "the recursion-depth sweep measured no cells; every run is "
-            "unavailable, and a report of those is not a curve"
-        )
-        logger.warning(
-            EVALS_RECURSION_NO_CELLS,
-            planned_cells=planned_cells,
-            recorded_cells=len(cells),
-        )
-        raise RecursionDepthNoCellsMeasuredError(msg)
-    required = provenance.requirement_count
-    return RecursionDepthReport(
-        provenance=provenance,
-        cells=tuple(cells),
-        by_achieved_depth=curve_by_achieved_depth(measured, requirement_count=required),
-        by_depth_cap=curve_by_depth_cap(measured, requirement_count=required),
-        achieved_depth_histogram=achieved_depth_histogram(measured),
-        caveats=tuple(caveats),
     )
